recommendation-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
